Split out common rest config schema code
ConfigValidationSchemaRESTProxy will work unchanged with mpOpenAPI-1.x
and 2.0, so split it out into a common bundle with all of its supporting
files.
