+++ dev/com.ibm.websphere.appserver.features/visibility/auto/com.ibm.websphere.appserver.configValidationConfigSchema-1.0.feature
@@ -6,6 +6,7 @@ IBM-Provision-Capability:\
   osgi.identity; filter:="(&(type=osgi.subsystem.feature)(|(osgi.identity=com.ibm.websphere.appserver.mpOpenAPI-1.0)(osgi.identity=com.ibm.websphere.appserver.mpOpenAPI-1.1)))"
 IBM-Install-Policy: when-satisfied
 -bundles=\
- com.ibm.ws.rest.handler.config.openapi
+ com.ibm.ws.rest.handler.config.openapi,\
+ io.openliberty.rest.handler.config.openapi.common
 kind=ga
 edition=core
