+++ dev/com.ibm.ws.rest.handler.config.openapi/bnd.bnd
@@ -1,5 +1,5 @@
 #*******************************************************************************
-# Copyright (c) 2019 IBM Corporation and others.
+# Copyright (c) 2019, 2021 IBM Corporation and others.
 # All rights reserved. This program and the accompanying materials
 # are made available under the terms of the Eclipse Public License v1.0
 # which accompanies this distribution, and is available at
@@ -19,15 +19,8 @@ Bundle-Description: Open API information for REST-based config resources; versio
 
 WS-TraceGroup: rest.config
 
-Web-ContextPath: openapi/platform
-OL-VirtualHost: ${admin.virtual.host}
-
-IBM-Authorization-Roles: com.ibm.ws.management
-IBM-Web-Extension-Processing-Disabled: true
-
 Include-Resource: \
-    META-INF=resources/META-INF,\
-    WEB-INF=resources/WEB-INF
+    META-INF=resources/META-INF
   
 Private-Package:\
   com.ibm.ws.rest.handler.config.openapi.*
@@ -37,11 +30,9 @@ Private-Package:\
 
 -buildpath:\
   com.ibm.websphere.appserver.spi.logging,\
-  com.ibm.websphere.javaee.servlet.3.0;version=latest,\
   com.ibm.websphere.org.osgi.core,\
   com.ibm.websphere.rest.handler;version=latest,\
   com.ibm.ws.org.osgi.annotation.versioning;version=latest,\
-  com.ibm.ws.rest.handler;version=latest,\
   com.ibm.wsspi.org.osgi.service.component.annotations;version=latest,\
   com.ibm.ws.microprofile.openapi;version=latest,\
   com.ibm.websphere.org.eclipse.microprofile.openapi.1.0;version=latest,\
