+++ dev/io.openliberty.rest.handler.config.openapi.common/.settings/bndtools.core.prefs
@@ -1,0 +1,2 @@
+compileErrorAction=build
+eclipse.preferences.version=1
