+++ dev/io.openliberty.rest.handler.config.openapi.common/resources/io/openliberty/rest/handler/config/openapi/common/resources/CWWKOMessages_fr.nlsprops
